페이지 컴포넌트 분리 (_컴포넌트.tsx) <- 요거 어떠십니까 예원님?
